+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.2] - 2023-12-08
 ### Changed
-- `@rarimo/rarime` - Update credentials supported hostnames
+- `@rarimo/rarime` - added new dashboard domains to the whitelist;
 
 ## [1.0.1] - 2023-11-29
 ### Fixed