+++ CHANGELOG.md
@@ -12,7 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `@rarimo/rarime` - credentials storage tests
 
 ### Changed
-- `@rarimo/rarime` - refactored credential-helpers methods to `VCManager` class
+- `@rarimo/rarime`
+  - refactored credential-helpers methods to `VCManager` class
+  - Interacting with ceramic ComposeDB by graphql requests
 
 ## [1.0.2] - 2023-12-08
 ### Changed