+++ CHANGELOG.md
@@ -14,7 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - `@rarimo/rarime`
   - refactored credential-helpers methods to `VCManager` class
-  - Interacting with ceramic ComposeDB by graphql requests
+    - ceramic account injection
+    - unified ceramic client
+    - hash and encrypt data
+  - Moved credentials to ComposeDB
 
 ## [1.0.2] - 2023-12-08
 ### Changed